% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tensor_regressions.R
\name{C1_reg}
\alias{C1_reg}
\title{Compute C1 regression}
\usage{
C1_reg(init_list, Y, X, R)
}
\arguments{
\item{init_list}{The initial list of CP decomposition factors.}

\item{Y}{The target tensor.}

\item{X}{The input tensor.}

\item{R}{The rank of the CP decomposition.}
}
\value{
The C1 regularization factor matrix.
}
\description{
This function computes the C1 regression term for the CP regression
algorithm.
}
\seealso{
\code{\link{C2_reg}}, \code{\link{D1_reg}}, \code{\link{D2_reg}}
\code{\link{cp_regression}}
}
