% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tensor_regressions.R
\name{C2_reg}
\alias{C2_reg}
\title{Compute C2 regression for CP regression}
\usage{
C2_reg(init_list, Y, X, R)
}
\arguments{
\item{init_list}{The initial list of CP decomposition factors.}

\item{Y}{The target tensor.}

\item{X}{The input tensor.}

\item{R}{The rank of the CP decomposition.}
}
\value{
The C2 regression factor matrix.
}
\description{
This function computes the C2 regression term for the CP regression algorithm.
It is used as part of the larger CP regression algorithm to estimate the factor
matrix associated with the first dimension of the target tensor.
}
\seealso{
\code{\link{C1_reg}}, \code{\link{D1_reg}}, \code{\link{D2_reg}}
\code{\link{cp_regression}}
}
