% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tensor_decompositions.R
\name{cp_modified}
\alias{cp_modified}
\title{Canonical Polyadic Decomposition}
\usage{
cp_modified(
  tnsr,
  num_components = NULL,
  max_iter = 25,
  tol = 1e-05,
  norm_type = "2"
)
}
\arguments{
\item{tnsr}{Tensor with K modes}

\item{num_components}{the number of rank-1 K-Tensors to use in approximation}

\item{max_iter}{maximum number of iterations if error stays above \code{tol}}

\item{tol}{relative Frobenius norm error tolerance}

\item{norm}{Type of norm to obtain lambdas. "O" for one norm, "I" 
for infinity norm, "F" for frobenius norm, "M" for maximum modulus, 
"2" for "spectral" or 2-norm}
}
\value{
a list containing the following \describe{
\item{\code{lambdas}}{a vector of normalizing constants, one for each component}
\item{\code{U}}{a list of matrices - one for each mode - each matrix with \code{num_components} columns}
\item{\code{conv}}{whether or not \code{resid} < \code{tol} by the last iteration}
\item{\code{norm_type}}{Change norm type for calculating lambdas.}
\item{\code{norm_percent}}{the percent of Frobenius norm explained by the approximation}
\item{\code{est}}{estimate of \code{tnsr} after compression}
\item{\code{fnorm_resid}}{the Frobenius norm of the error \code{fnorm(est-tnsr)}}
\item{\code{all_resids}}{vector containing the Frobenius norm of error for all the iterations}
}
}
\description{
Modified version of Canonical Polyadic decomposition. Here, you can
specify the type of norm to take for the lambdas, with Euclidean
norm being the default
}
\details{
Uses the Alternating Least Squares (ALS) estimation procedure. A progress bar is included to help monitor operations on large tensors.
}
\references{
T. Kolda, B. Bader, "Tensor decomposition and applications". SIAM Applied Mathematics and Applications 2009.
}
\seealso{
\code{\link{tucker}}
}
