% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tensor_regressions.R
\name{cp_regression}
\alias{cp_regression}
\title{Perform CP regression}
\usage{
cp_regression(
  Y,
  X,
  R,
  obs_dim_X,
  obs_dim_Y,
  convThresh = 1e-05,
  max_iter = 500,
  seed = 0
)
}
\arguments{
\item{Y}{The input tensor.}

\item{X}{The covariate tensor.}

\item{R}{The rank of the CP decomposition.}

\item{obs_dim_X}{The observed dimensions of the covariate tensor X.}

\item{obs_dim_Y}{The observed dimensions of the input tensor Y.}

\item{convThresh}{The convergence threshold for the CP regression algorithm.}

\item{max_iter}{The maximum number of iterations.}

\item{seed}{The random seed for reproducibility.}
}
\value{
A list containing the estimated tensor B, factor matrices, 
convergence status, and the number of iterations.
}
\description{
This function performs CP regression to estimate the tensor B given
the input tensor Y and covariate tensor X.
}
\examples{
Y <- # input tensor
X <- # covariate tensor
R <- 3
obs_dim_X <- # observed dimensions of X
obs_dim_Y <- # observed dimensions of Y
cp_regression(Y, X, R, obs_dim_X, obs_dim_Y, convThresh = 1e-05,
              max_iter = 500, seed = 0)

}
\seealso{
\code{\link{y_regression}}, \code{\link{x_regression}} 
\code{\link{conv_cond}}
}
