% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tensor_decompositions.R
\name{cp_uniqueness}
\alias{cp_uniqueness}
\title{Check for Uniqueness of CP Decomposition}
\usage{
cp_uniqueness(cp_object)
}
\arguments{
\item{cp_object}{A CP decomposition object from the 'cptensor' package.}
}
\value{
A list with boolean values for the sufficient and necessary conditions
}
\description{
Determines whether a CP decomposition is unique based on sufficient and
necessary conditions.
}
\details{
The function checks for uniqueness of a CP decomposition based on
sufficient and necessary conditions. The sufficient condition checks if the
sum of the ranks of each factor matrix is greater than or equal to 2 times
the rank of the tensor plus the number of modes minus one. The necessary
condition checks if the minimum product of ranks of each factor matrix is
greater than or equal to the rank of the tensor.
}
\examples{
library(rTensor)
X <- rand_tensor(c(4,5,6))
fit <- cp(X, 3)
cp_uniqueness(fit)

}
