% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tensor_regressions.R
\name{D1_reg}
\alias{D1_reg}
\title{Compute D1 regression for CP regression}
\usage{
D1_reg(init_list, Y, X, R)
}
\arguments{
\item{init_list}{The initial list of CP decomposition factors.}

\item{Y}{The response tensor}

\item{X}{The predictor tensor}

\item{R}{The CP rank}
}
\value{
The D1 regression factor matrix.
}
\description{
This function computes the D1 regression term for the CP regression algorithm.
It is used as part of the larger CP regression algorithm to estimate the factor
matrix associated with the third dimension of the target tensor.
}
\seealso{
\code{\link{C1_reg}}, \code{\link{C2_reg}}, \code{\link{D2_reg}}
\code{\link{cp_regression}}
}
