% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tensor_regressions.R
\name{D2_reg}
\alias{D2_reg}
\title{Compute D2 regression for CP regression}
\usage{
D2_reg(init_list, X, Y, R)
}
\arguments{
\item{init_list}{The initial list of CP decomposition factors.}

\item{X}{The tensor X used in the CP regression.}

\item{Y}{The target tensor Y used in the CP regression.}

\item{R}{The rank of the CP decomposition.}
}
\value{
The D2 regression factor matrix.
}
\description{
This function computes the D2 regression term for the CP regression algorithm.
It is used as part of the larger CP regression algorithm to estimate the factor
matrix associated with the fourth dimension of the target tensor.
}
\seealso{
\code{\link{C1_reg}}, \code{\link{C2_reg}}, \code{\link{D1_reg}}
\code{\link{cp_regression}}
}
