% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tensor_regressions.R
\name{HOOLS}
\alias{HOOLS}
\title{Compute the Higher-Order Ordinary Least Squares (HOOLS) estimator}
\usage{
HOOLS(Y, X)
}
\arguments{
\item{Y}{A tensor of response variables.}

\item{X}{A tensor of predictor variables.}
}
\value{
A tensor of HOOLS estimators.
}
\description{
The HOOLS estimator is a generalization of the ordinary least squares (OLS) estimator to higher-order tensors.
}
\details{
This function computes the HOOLS estimator for a tensor regression model. The HOOLS estimator is defined
as the solution to the equation X = X_1 B_1 + X_2 B_2 + ... + X_r B_r, where X is a tensor of predictor variables,
X_1, X_2, ..., X_r are subtensors of X obtained by fixing the values of certain indices, and B_1, B_2, ..., B_r
are tensors of HOOLS estimators.
}
\examples{
# Generate random tensor data
library(TensorEconometrics)
Y <- rand_tensor(c(100, 3,4))
X <- rand_tensor(c(100, 3,2))

# Compute HOOLS estimator with observations along the first dimension
HOOLS(Y, X, obs_dim_Y = 1, obs_dim_X = 1)

}
