% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tensor_decompositions.R
\name{reconstruct_cp}
\alias{reconstruct_cp}
\title{Reconstruct CP tensor from factor matrices}
\usage{
reconstruct_cp(A, B, C, D, r, lambda = rep(1, r))
}
\arguments{
\item{A}{factor matrix of mode 1}

\item{B}{factor matrix of mode 2}

\item{C}{factor matrix of mode 3}

\item{D}{factor matrix of mode 4}

\item{r}{rank of the CP decomposition}

\item{lambda}{vector of CP decomposition coefficients, defaults to vector of 1's}
}
\value{
Reconstructed CP tensor
}
\description{
Reconstructs the original CP tensor from its factor matrices, i.e., given the factor matrices
A, B, and C (corresponding to the three modes of the tensor) and a vector of CP decomposition
coefficients lambda, it computes the tensor T as the sum of the outer products of the columns of
the factor matrices scaled by the corresponding coefficient.
}
