% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tensor_algebra.R
\name{rnorm_tnsr}
\alias{rnorm_tnsr}
\title{Random tensor with standard deviation}
\usage{
rnorm_tnsr(modes = c(3, 4, 5), sd = 1, drop = FALSE)
}
\arguments{
\item{modes}{A vector specifying the dimensions of the tensor.}

\item{sd}{The standard deviation used to generate random values.}

\item{drop}{Whether or not modes equal to 1 should be dropped.}
}
\value{
A random tensor with the specified dimensions.
}
\description{
This function generates a random tensor with specified modes and an option
for setting the standard deviation.
}
