% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tensor_algebra.R
\name{spark}
\alias{spark}
\title{Calculate the Spark of a Given Matrix}
\usage{
spark(Phi)
}
\arguments{
\item{Phi}{a matrix}
}
\value{
A list with two elements: \code{K}, an integer specifying the 
  Spark of the matrix, and \code{columns}, a numeric vector of indices 
  specifying the columns that make up the linearly dependent set.
}
\description{
The Spark of a matrix is the smallest number of linearly dependent columns. 
This function takes a matrix \code{Phi} as input and returns the Spark of 
the matrix, \code{K}, and the indices of the columns that make up the 
linearly dependent set, \code{columns}.
}
\details{
This function iteratively chooses subsets of columns of \code{Phi} 
and checks if they are linearly dependent. It stops when it finds the 
smallest set of linearly dependent columns. If all subsets of columns are 
linearly independent, it returns the rank of the matrix plus one and all 
columns of the matrix.
}
\examples{
Phi <- matrix(c(1, 2, 3, 2, 4, 6, 3, 6, 9), nrow = 3, ncol = 3)
spark(Phi)

}
