% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{tensor_data}
\alias{tensor_data}
\title{GVAR Tensor Data}
\format{
An array with 161 observations, 32 countries, and 3 economic variables
}
\source{
https://www.mohaddes.org/gvar.
}
\usage{
tensor_data
}
\description{
Global VAR data in tensor form. Data format is a 161x32x3 array with 161
observations, 32 countries, and 3 economic indicators per country.
}
\keyword{datasets}
