% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tensor_regressions.R
\name{tensor_inverse}
\alias{tensor_inverse}
\title{Invert a tensor}
\usage{
tensor_inverse(A)
}
\arguments{
\item{A}{a tensor to be inverted}
}
\value{
the inverse of A
}
\description{
Given a tensor A, this function returns the inverse of A, where the inverse is
defined as the inverse of the matricization of A along the first half of its
dimensions.
}
\examples{
A <- array(1:8, dim = c(2, 2, 2))
tensor_inverse(A) # returns the inverse of A
}
