% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tensor_algebra.R
\name{tlag}
\alias{tlag}
\title{Tensor Lag Operator}
\usage{
tlag(ten_data, p = 1)
}
\arguments{
\item{ten_data}{Data in tensor format}

\item{p}{Number of lags}
}
\value{
A list with the adjusted original matrix and the lagged tensor
}
\description{
Creates lags for a tensor of size N_1 x N_2 x T
}
