% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{traditional_data}
\alias{traditional_data}
\title{Global Vector Autoregression Data Set}
\format{
A time series data frame with 120 observations of 15 variables:
  \describe{
    \item{USA_GDP}{Gross Domestic Product (GDP) for the United States}
    \item{USA_CPI}{Consumer Price Index (CPI) for the United States}
    \item{USA_IR}{Interest rate for the United States}
    \item{CAN_GDP}{Gross Domestic Product (GDP) for Canada}
    \item{CAN_CPI}{Consumer Price Index (CPI) for Canada}
    \item{CAN_IR}{Interest rate for Canada}
    \item{UK_GDP}{Gross Domestic Product (GDP) for the United Kingdom}
    \item{UK_CPI}{Consumer Price Index (CPI) for the United Kingdom}
    \item{UK_IR}{Interest rate for the United Kingdom}
    \item{GER_GDP}{Gross Domestic Product (GDP) for Germany}
    \item{GER_CPI}{Consumer Price Index (CPI) for Germany}
    \item{GER_IR}{Interest rate for Germany}
    \item{JPN_GDP}{Gross Domestic Product (GDP) for Japan}
    \item{JPN_CPI}{Consumer Price Index (CPI) for Japan}
    \item{JPN_IR}{Interest rate for Japan}
  }
}
\source{
This data set is simulated for demonstration purposes.
}
\usage{
traditional_data
}
\description{
This data set contains a global vector autoregression (VAR) model with 5
countries and 3 economic variables. Each column corresponds to a specific
country and economic variable combination.
}
\examples{
data(global_var)
head(global_var)

}
\keyword{data}
