% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tensor_algebra.R
\name{ttt}
\alias{ttt}
\title{Compute the tensor product of two tensors along specified modes.}
\usage{
ttt(A, B, alongA = NA, alongB = NA)
}
\arguments{
\item{A}{First input tensor.}

\item{B}{Second input tensor.}

\item{alongA}{An integer indicating the mode of A along which to perform
the tensor product. Default is NA.}

\item{alongB}{An integer indicating the mode of B along which to perform
the tensor product. If NA, set to alongA. Default is NA.}
}
\value{
A new tensor resulting from the tensor product of A and B along the
  specified modes.
}
\description{
Compute the tensor product of two tensors along specified modes.
}
\examples{
A <- array(1:24, dim = c(3, 4, 2))
B <- array(1:24, dim = c(2, 3, 4))
ttt(A, B, alongA = 3, alongB = 1)

}
