% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tensor_algebra.R
\name{ttv}
\alias{ttv}
\title{Tensor times Vector}
\usage{
ttv(A, V, N)
}
\arguments{
\item{A}{A tensor}

\item{V}{A column vector}

\item{N}{Dimension over which to multiply}
}
\value{
A contracted tensor
}
\description{
Given a tensor and a (column) vector, compute the product. The integer N 
specifies the dimension in A along which V is multiplied.
}
