% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tensor_decompositions.R
\name{tucker_rebuild}
\alias{tucker_rebuild}
\title{Rebuilds a Tucker tensor from a list of component matrices and core tensor}
\usage{
tucker_rebuild(comp_list)
}
\arguments{
\item{comp_list}{A list with the core tensor first then component matrices.}
}
\value{
The rebuilt Tucker tensor
}
\description{
This function takes a list of component matrices and rebuilds a Tucker tensor
by iteratively applying the ttm function.
}
