% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tensor_regressions.R
\name{tucker_regression}
\alias{tucker_regression}
\title{Perform Tucker regression}
\usage{
tucker_regression(Y, X, R, convThresh = 1e-04, max_iter = 400, seed = 0)
}
\arguments{
\item{Y}{The target tensor.}

\item{X}{The predictor tensor.}

\item{R}{A vector specifying the desired rank of each mode in the Tucker
decomposition.}

\item{convThresh}{The convergence threshold for the Tucker decomposition.}

\item{max_iter}{The maximum number of iterations for convergence.}

\item{seed}{An optional seed for reproducible results.}
}
\value{
A list with: G (core tensor), U[[i]] is a list of factor matrices, est is estimate for full coef, num_iters is number of iterations until convergence, converged is bool (whether or not converged), mse is mean squared error
}
\description{
This function performs Tucker regression on a given tensor, decomposing it
into a low-rank tensor and factor matrices.
}
