% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tensor_regressions.R
\name{tucker_regression2}
\alias{tucker_regression2}
\title{Alternate Tucker Regression}
\usage{
tucker_regression2(Y, X, R, convThresh = 1e-05, max_iter = 400, seed = 0)
}
\arguments{
\item{Y}{The target tensor.}

\item{X}{The predictor tensor.}

\item{R}{A vector specifying the desired rank of each mode in the Tucker
decomposition.}

\item{convThresh}{The convergence threshold for the Tucker decomposition.}

\item{max_iter}{The maximum number of iterations for convergence.}

\item{seed}{An optional seed for reproducible results.}
}
\value{
A list containing the decomposed components and the rebuilt tensor.
}
\description{
This function is similar to the original Tucker Regression, but allows for
observations in the first dimension rather than the last dimension. This is
in line with Brandi et. al 2021.
}
