% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tensor_regressions.R
\name{x_regression}
\alias{x_regression}
\title{Regression associated with X}
\usage{
x_regression(init_list, Y, X, R, idx)
}
\arguments{
\item{init_list}{The initial list of CP decomposition factors.}

\item{Y}{The target tensor.}

\item{X}{The input tensor.}

\item{R}{The rank of the CP decomposition.}

\item{idx}{The index associated with the X regression}
}
\value{
The factor matrix associated with the X regression
}
\description{
This function computes the regression for the X portion of the CP 
regression. This is part one of a two part scheme for the parameters
of the CP regression.
}
\seealso{
\code{\link{y_regression}}, \code{\link{conv_cond}}
\code{\link{cp_regression}}
}
