% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tensor_regressions.R
\name{y_regression}
\alias{y_regression}
\title{Regression associated with Y}
\usage{
y_regression(init_list, Y, X, R, Ddims, idx)
}
\arguments{
\item{init_list}{The initial list of CP decomposition factors.}

\item{Y}{The response tensor}

\item{X}{The predictor tensor}

\item{R}{The CP rank}

\item{Ddims}{Vector with dimensions of D}

\item{idx}{Index for Y tensor associated with CP regression}
}
\value{
The D1 regression factor matrix.
}
\description{
This function computes the regression for the Y portion of the CP 
regression. This is part two of a two part scheme for the parameters
of the CP regression.
}
\seealso{
\code{\link{x_regression}}, \code{\link{conv_cond}}
\code{\link{cp_regression}}
}
