% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tensor_regressions.R
\name{matrix_ols_sim}
\alias{matrix_ols_sim}
\title{Higher Order Ordinary Least Squares (HOOLS) Estimator}
\usage{
matrix_ols_sim(obs, parameter_matrix, num_rows, num_cols)
}
\arguments{
\item{Y}{A higher order array with dimensions `d1 x d2 x ... x dn`.}

\item{X}{A higher order array with dimensions `d1 x d2 x ... x dn`.}

\item{obs_dim_Y}{An integer indicating the dimension containing the
observations in `Y` (default is `n`).}

\item{obs_dim_X}{An integer indicating the dimension containing the
observations in `X` (default is `n`).}
}
\value{
Returns the OLS estimator.
}
\description{
Calculates the Ordinary Least Squares (OLS) Estimator for a
higher order array. The last dimension of `Y` and `X` is assumed to contain
the observations by default. IMPORTANT: Assumes that both response and 
predictor tensors are of the same dimensions.
}
\examples{
HOOLS(Y, X)

}
