% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{traditional_data}
\alias{traditional_data}
\title{GVAR Traditional Data}
\format{
: A matrix with 161 observations and 96 variables. 
 
\describe{
 \item{y.ARGENTINA}{Argentina GDP}
 \item{Dp.ARGENTINA}{Argentina Inflation Rate}
 \item{r.ARGENTINA}{Argentina Interest Rate}
}
}
\source{
: <https://www.mohaddes.org/gvar>
}
\usage{
traditional_data
}
\description{
A flattened version of the Global VAR data. Captures GDP, inflation, and 
interest rate of 32 countries. For more information, go to
https://www.mohaddes.org/gvar.
}
\keyword{datasets}
