% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tensor_regressions.R
\name{tucker_regression}
\alias{tucker_regression}
\title{Perform Tucker regression}
\usage{
tucker_regression(
  Y,
  X,
  R,
  obs_dim_X,
  obs_dim_Y,
  convThresh = 1e-05,
  max_iter = 400,
  seed = 0
)
}
\arguments{
\item{Y}{The target tensor.}

\item{X}{The predictor tensor.}

\item{R}{A vector specifying the desired rank of each mode in the Tucker
decomposition.}

\item{obs_dim_X}{A vector specifying the observed dimensions of the predictor
tensor.}

\item{obs_dim_Y}{A vector specifying the observed dimensions of the target
tensor.}

\item{convThresh}{The convergence threshold for the Tucker decomposition.}

\item{max_iter}{The maximum number of iterations for convergence.}

\item{seed}{An optional seed for reproducible results.}
}
\value{
A list containing the decomposed components and the rebuilt tensor.
}
\description{
This function performs Tucker regression on a given tensor, decomposing it
into a low-rank tensor and factor matrices.
}
