% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tensor_regressions.R
\name{tucker_regression}
\alias{tucker_regression}
\title{Perform Tucker regression}
\usage{
tucker_regression(X, Y, R, convThresh = 1e-04, max_iter = 400, seed = 0)
}
\arguments{
\item{X}{The predictor tensor.}

\item{Y}{The target tensor.}

\item{R}{A vector specifying the desired rank of each mode in the Tucker
decomposition.}

\item{convThresh}{The convergence threshold for the Tucker decomposition.}

\item{max_iter}{The maximum number of iterations for convergence.}

\item{seed}{An optional seed for reproducible results.}
}
\value{
A list containing the decomposed components and the rebuilt tensor.
}
\description{
This function performs Tucker regression on a given tensor, decomposing it
into a low-rank tensor and factor matrices.
}
